requirements.txt
nltk==3.8.1
scikit-learn==1.6.1
numpy==1.26.4
streamlit==1.31.0
chardet==5.2.0
spacy==3.7.2
PyPDF2==3.0.1
python-docx==1.0.1

# Download spaCy model
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl
